refactor(cli): extract trash command module
